Update : PathFind内での探査で隣接ノードのみの探査になっていなかったので、GetNeighbor関数を追加して、隣接ノードのみ調べるように調整

+++ ForStudySesssion/Assets/Scripts/Path/Node/Node.cs
@@ -2,11 +2,11 @@ using UnityEngine;
 
 public class Node : MonoBehaviour
 {
-    [HideInInspector] public float X, Y;
+    [HideInInspector] public int X, Y;
     [HideInInspector] public bool Movable;
     [HideInInspector] public float Cost;
 
-    public void Init(float X, float Y, bool Movable = true, float Cost = 1)
+    public void Init(int X, int Y, bool Movable = true, float Cost = 1)
     {
         this.X = X;
         this.Y = Y;
